AI: refactor the data process module into DLSA and vision for disease prediction
