+++ applications/ai/disease_prediction/src/disease_prediction/dlsa/trainer.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions
 # and limitations under the License.
 #
-import os
 
 from datasets import load_dataset, Features, Value, ClassLabel
