+++ applications/ai/disease_prediction/src/disease_prediction/vision/data/process.py
