+++ applications/ai/disease_prediction/src/disease_prediction/vision/data/split.py
@@ -1,31 +1,12 @@
 import os
 import shutil
 from os import path, listdir
-from pandas import DataFrame, read_csv
-from typing import Tuple
-from sklearn.model_selection import train_test_split
+from pandas import DataFrame
 from pathlib import Path
 
 
-def get_vision_split_output_dir(output_dir):
+def get_split_output_dir(output_dir):
     return os.path.join(output_dir, "split_images")
-
-
-def get_dlsa_split_output_dir(output_dir):
-    return os.path.join(output_dir, "annotation", "split")
-
-
-def split_dlsa_data(df: DataFrame, test_size: float) -> Tuple[DataFrame, DataFrame]:
-    """Split the dataset into training and testing sets for NLP.
-
-    Args:
-        df: Pandas DataFrame containing the data.
-        test_size: Proportion of the dataset to include in the test split.
-
-    Returns:
-        Tuple of Pandas DataFrames for training and testing, respectively.
-    """
-    return train_test_split(df, test_size=test_size)
 
 
 def get_subject_id(image_name):
@@ -63,9 +44,9 @@ def copy_images(patient_ids: DataFrame, source_folder: str, target_folder: str) 
             shutil.copy(full_src_path, target_folder)
 
 
-def split_vision_data(
+def split_data(
         train_data: DataFrame, test_data: DataFrame,
-        processed_data_dir, dataset_config, output_dir) -> dict:
+        processed_data_dir, dataset_config, output_dir):
     """
     This function creates and organizes the train and test dataset for vision task.
 
@@ -86,7 +67,7 @@ def split_vision_data(
     label_column = dataset_config["features"]["class_label"]
 
     # Create the target directory for the dataset
-    split_images_dir = get_vision_split_output_dir(output_dir)
+    split_images_dir = get_split_output_dir(output_dir)
     if path.exists(split_images_dir):
         shutil.rmtree(split_images_dir)
 
@@ -108,39 +89,16 @@ def split_vision_data(
             patient_ids = df[df[label_column] == label]
             copy_images(patient_ids, source_folder, target_folder)
 
+    return split_images_dir
+
 
 def split(
-        processed_data_dir, output_dir,
-        test_size, dataset_config, overwrite=True):
-    output_annotations_file = os.path.join(
-        os.path.join(processed_data_dir, "annotation"), "annotation.csv")
-
-    split_data_path = get_dlsa_split_output_dir(output_dir)
-    os.makedirs(split_data_path, exist_ok=True)
-
-    training_data_path = os.path.join(split_data_path, "train.csv")
-    testing_data_path = os.path.join(split_data_path, "test.csv")
-
-    if (path.exists(training_data_path) or path.exists(
-            testing_data_path)) and not overwrite:
-        print("Training or testing data already exist. Skip split.")
-        return
-
-    # create training and testing data for NLP if overwrite_training_testing_ids is True
-    # read the input data
-    input_data = read_csv(output_annotations_file)
-
-    # create training and testing data
-    training_data, testing_data = split_dlsa_data(
-        input_data, test_size)
-
-    # save training and testing data
-    training_data.to_csv(training_data_path, index=False)
-    testing_data.to_csv(testing_data_path, index=False)
-
+        dlsa_training_data, dlsa_testing_data,
+        processed_data_dir, dataset_config,
+        output_dir):
     # create vision data
-    split_vision_data(
-        training_data, testing_data,
+    split_data(
+        dlsa_training_data, dlsa_testing_data,
         processed_data_dir=processed_data_dir,
         dataset_config=dataset_config,
         output_dir=output_dir)
